+++ src/main/resources/ddl/initdb.sql
@@ -5,7 +5,7 @@ CREATE TABLE account
     password varchar (60) NOT NULL,
     email varchar (255),
     authority varchar(10) NOT NULL,
-    balance int NOT NULL DEFAULT 0,
+    balance decimal(11,2) NOT NULL DEFAULT 0,
     createdt timestamp NOT NULL,
     enabled boolean NOT NULL,
     expiredt date,
